requirements.txt
Django==5.1
Channels==4.0.0
Channels-redis==4.1.0
Daphne==4.0.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3